+++ lib/webpacker/dev_server_runner.rb
@@ -8,11 +8,13 @@ module Webpacker
   class DevServerRunner < Webpacker::Runner
     def run
       load_config
+      detect_unsupported_switches!
       detect_port!
       execute_cmd
     end
 
     private
+
       def load_config
         app_root = Pathname.new(@app_path)
 
@@ -28,11 +30,27 @@ module Webpacker
         @port              = dev_server.port
         @pretty            = dev_server.pretty?
         @hmr               = dev_server.hmr?
+        @https             = dev_server.https?
 
       rescue Errno::ENOENT, NoMethodError
         $stdout.puts "webpack dev_server configuration not found in #{@config.config_path}[#{ENV["RAILS_ENV"]}]."
         $stdout.puts "Please run bundle exec rails webpacker:install to install Webpacker"
         exit!
+      end
+
+      UNSUPPORTED_SWITCHES = %w[--host --port]
+      private_constant :UNSUPPORTED_SWITCHES
+      def detect_unsupported_switches!
+        unsupported_switches = UNSUPPORTED_SWITCHES & @argv
+        if unsupported_switches.any?
+          $stdout.puts "The following CLI switches are not supported by Webpacker: #{unsupported_switches.join(' ')}. Please edit your command and try again."
+          exit!
+        end
+
+        if @argv.include?("--https") && !@https
+          $stdout.puts "Please set https: true in webpacker.yml to use the --https command line flag."
+          exit!
+        end
       end
 
       def detect_port!
@@ -61,6 +79,7 @@ module Webpacker
         cmd += ["--config", @webpack_config]
         cmd += ["--progress", "--color"] if @pretty
         cmd += ["--hot"] if @hmr
+        cmd += @argv
 
         Dir.chdir(@app_path) do
           Kernel.exec env, *cmd