+++ test/mounted_app/test/dummy/config/application.rb
@@ -1,0 +1,10 @@
+require "action_controller/railtie"
+require "action_view/railtie"
+require "webpacker"
+
+module TestDummyApp
+  class Application < Rails::Application
+    config.secret_key_base = "abcdef"
+    config.eager_load = true
+  end
+end